chore: Some restructuring in `dash-spv::sync`
