+++ dash-spv/src/sync/headers/mod.rs
@@ -1,0 +1,5 @@
+//! Header synchronization with fork detection and reorganization handling.
+
+mod manager;
+
+pub use manager::{HeaderSyncManager, ReorgConfig};
