+++ dash-spv/src/sync/headers2/mod.rs
@@ -1,0 +1,5 @@
+//! Headers2 state management for compressed header synchronization.
+
+mod manager;
+
+pub use manager::{Headers2StateManager, Headers2Stats, ProcessError};
