+++ dash-spv/src/sync/masternodes/mod.rs
@@ -1,0 +1,6 @@
+//! Masternode synchronization and embedded data.
+
+pub mod embedded_data;
+mod manager;
+
+pub use manager::MasternodeSyncManager;
